% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_random_experiments.R
\name{complex_random_experiments}
\alias{complex_random_experiments}
\title{Run K random experiments}
\usage{
complex_random_experiments(
  X,
  y,
  K = 20,
  T_stop = 1,
  num_dummies = ncol(X),
  dummy_type = c("Complex Gaussian", "Complex Circular Uniform")[1],
  method = c("ctrex", "ctrex+GVS")[1],
  gvs_type = c("cEN", "cIEN")[1],
  hc_method = "ward.D2",
  tree_cut_type = c("fixed", "dynamic")[1],
  coherence_max = 0.5,
  minModuleSize = 10,
  lambda_2_lars = NULL,
  early_stop = TRUE,
  ctlars_learner_lst = NULL,
  verbose = TRUE,
  intercept = FALSE,
  standardize = TRUE,
  parallel_process = FALSE,
  parallel_max_cores = min(K, max(1, parallel::detectCores(logical = FALSE))),
  seed = NULL,
  eps = .Machine$double.eps
)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{y}{Response vector.}

\item{K}{Number of random experiments.}

\item{T_stop}{Number of included dummies after which the random experiments (i.e., forward selection processes) are stopped.}

\item{num_dummies}{Number of dummies that are appended to the predictor matrix.}

\item{method}{'ctrex' for the T-Rex selector. 'ctrex+GVS' for CT-Rex Group-Variable Selector.}

\item{gvs_type}{'cEN' for elastic network group selection. 'cIEN' for informed-elastic network group selection.}

\item{early_stop}{Logical. If TRUE, then the forward selection process is stopped after T_stop dummies have been included.
Otherwise the entire solution path is computed.}

\item{ctlars_learner_lst}{A list of ctlars learners.}

\item{verbose}{Logical. If TRUE progress in computations is shown.}

\item{intercept}{Default FALSE.}

\item{standardize}{Logical. If TRUE the predictors are standardized and the response is centered.}

\item{parallel_process}{Logical. If TRUE random experiments are executed in parallel.}

\item{parallel_max_cores}{Maximum number of cores to be used for parallel processing
(default: minimum{Number of random experiments K, number of physical cores}).}

\item{seed}{Seed for random number generator (ignored if parallel_process = FALSE).}

\item{eps}{Numerical zero.}
}
\value{
List containing the results of the K random experiments.
}
\description{
Run K random experiments and compute the matrix of relative occurrences for all variables
and all numbers of included variables before stopping.
}
\examples{
set.seed(123)
}
