% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ctrex_utils.R
\docType{package}
\name{CTrex-package}
\alias{CTrex}
\alias{CTrex-package}
\title{CTrex: ctrex - Complex Terminating Random Experiment Selector}
\description{
The package implements a complex valued T-Rex selector.
}
\details{
ctrex is a package showing how to produce packages with
Rcpp Modules and C++ classes and make them available under S4.
}
\author{
\strong{Maintainer}: Fabian Scheidt \email{fabian.scheidt@tu-darmstadt.de}

Authors:
\itemize{
  \item Jasin Machkour \email{jasin.machkour@tu-darmstadt.de}
  \item Michael Muma \email{michael.muma@tu-darmstadt.de}
}

}
