% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrex.R
\name{ctrex}
\alias{ctrex}
\title{Run the CT-Rex selector}
\usage{
ctrex(
  X,
  y,
  tFDR = 0.2,
  K = 20,
  max_num_dummies = 10,
  max_T_stop = TRUE,
  method = c("ctrex", "ctrex+GVS")[1],
  gvs_type = c("cEN", "cIEN")[1],
  dummy_type = c("Complex Gaussian", "Complex Circular Uniform")[1],
  hc_method = c("single", "complete", "ward.D2")[1],
  coherence_max = 0.5,
  parallel_process = FALSE,
  parallel_max_cores = min(K, max(1, parallel::detectCores(logical = FALSE))),
  seed = NULL,
  eps = .Machine$double.eps,
  verbose = TRUE
)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{y}{Response vector.}

\item{tFDR}{Target FDR level (between 0 and 1, i.e., 0\% and 100\%).}

\item{K}{Number of random experiments.}

\item{max_num_dummies}{Integer factor determining the maximum number of dummies as a multiple of the number of original variables p
(i.e., num_dummies = max_num_dummies * p).}

\item{max_T_stop}{If TRUE the maximum number of dummies that can be included before stopping is set to ceiling(n / 2),
where n is the number of data points/observations.}

\item{method}{'trex' for the T-Rex selector. 'trex+GVS' for CT-Rex Group-Variable Selector.}

\item{gvs_type}{'EN' for elastic network group selection. 'IEN' for informed-elastic network group selection.}

\item{parallel_process}{Logical. If TRUE random experiments are executed in parallel.}

\item{parallel_max_cores}{Maximum number of cores to be used for parallel processing
(default: minimum{Number of random experiments K, number of physical cores}).}

\item{seed}{Seed for random number generator (ignored if parallel_process = FALSE).}

\item{eps}{Numerical zero.}

\item{verbose}{Logical. If TRUE progress in computations is shown.}
}
\value{
A list containing the estimated support vector and additional information, including the number of used dummies and the number of included dummies before stopping.
}
\description{
The CT-Rex selector performs fast variable selection in high-dimensional settings while
controlling the false discovery rate (FDR) at a user-defined target level.
}
\examples{
set.seed(1)
}
